final script for novartispdx and some changes to calc_pdx_metrics.py
Changes to calc_pdx_metrics.py include adding flexibility for `study` and `source` instead of being hard coded to MPNST. Added an if statement around the final touches to combos dataset, because if there is no drug data for the combo drugs, errors arise. Moved writing combos to csv to be inside this if statement.